Initial reorganization.
