+++ controllers/lrgs-controller/Cargo.toml
@@ -37,3 +37,10 @@ jsonptr = "0.7.1"
 anyhow = "1.0.97"
 lazy_static = "1.4.0"
 passwords = "3.1.16"
+actix-web = "4.10.2"
+tracing = "0.1.41"
+tracing-subscriber = { version = "0.3.19", features = ["json", "env-filter"] }
+tracing-opentelemetry = "0.28.0"
+opentelemetry = { version = "0.27.1", features = ["trace"] }
+opentelemetry-otlp = { version = "0.27.0", optional = true }
+opentelemetry_sdk = { version = "0.27.1", features = ["rt-tokio"] }
