+++ controllers/lrgs-controller/src/lrgs/mod.rs
@@ -3,3 +3,6 @@ pub mod password_file;
 pub mod statefulset;
 pub mod configmap;
 pub mod service;
+pub mod controller;
+pub mod state;
+pub mod telemetry;
