+++ controllers/lrgs-controller/src/lrgs/telemetry.rs
@@ -1,0 +1,5 @@
+
+
+pub async fn init() {
+
+}
